requirements.txt
streamlit>=1.28.0
streamlit-webrtc>=0.47.0
opencv-python-headless>=4.8.0
cvzone>=1.5.6
google-generativeai>=0.3.0
pillow>=9.5.0
numpy>=1.24.0
av>=10.0.0
mediapipe>=0.10.0